fix(loginEvent): chore: refactor some common utility methods
